test-requirements.txt
apache-airflow==1.10.11
codecov==2.1.8
marquez-python==0.5.6
pytest==6.0.1
pytest-cov==2.10.1
requests==2.24.0
psycopg2-binary==2.8.5